practica 1

+++ practica1_2.0/ejercicio5.py
@@ -1,9 +1,8 @@
-#! /usr/bin/python
+#! /usr/bin/python3.6
 import os
 
 directorio = os.getcwd()
 print('El directorio actual de trabajo es: ' + directorio)
-
 
 print('Los ficheros que hay en ' + directorio + ' son:')
 ficheros = os.listdir('.')
@@ -13,10 +12,14 @@ for x in ficheros:
 print('Escriba el nombre del fichero que desea renombrar:')
 fichero = input()
 
-if os.path.isfile(fichero):
+if os.path.isfile(fichero): 
 	print('Escriba el nuevo nombre del fichero:')
 	renombre = input()
-	os.rename(fichero, renombre)
-	print('El fichero ha sido renombrado correctamente')
-else:
-	print('El fichero que desea renombrar no existe')
+	if os.path.isfile(renombre):
+		print('Imposible renombrar. El nuevo nombre del fichero ya existe.')
+	else:
+		os.rename(fichero, renombre)
+		print('El fichero ha sido renombrado correctamente.')
+		
+else: 
+	print('El fichero que desea renombrar no existe.')
